statistic added calendar as dialog
