+++ lib/view/statisticView.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class StatisticView extends StatefulWidget {
+  final Function(Widget) callback;
+
+  StatisticView(this.callback);
+
+  @override
+  State<StatefulWidget> createState() => _StatisticViewState();
+}
+
+class _StatisticViewState extends State<StatisticView> {
+  @override
+  Widget build(BuildContext context) {
+    return null;
+  }
+}
